requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
aiosqlite
python-multipart
pydantic
python-docx
jinja2
boto3
redis
rq
pdfminer.six
ocrmypdf
PyMuPDF
authlib
python-dotenv
clamd
requests
prometheus-client
itsdangerous

httpx

pydantic-settings
python-jose[cryptography]

pdfminer.six==20231228

redis==5.0.7
